requirements.txt
Django==4.2.7
Pillow==10.1.0
qrcode==7.4.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
whitenoise==6.6.0
django-colorfield==0.10.1
djangorestframework==3.14.0
